Improve defensive copying
Make a defensive copy of MUXSDKCustomerData when it is provided. Since
it is a reference, this prevents alterations the SDK makes to bubble
up to client code. As well as guards against the player software name
and version from being altered after they're set.

Improve readability.

Add customer data tests to validate dimensions are passed through and
check for pointer equality